requirements.txt
fastapi
uvicorn
langgraph
langchain
langchain-openai
langchain-chroma
chromadb
openai
requests
pydantic
python-dotenv
sqlalchemy
pymysql
cryptography